add historical quotes support

+++ src/Gateways/QuotesGateway/Models/HistoryQuoteInfo.cs
@@ -8,10 +8,10 @@ namespace InvestipsApiContainers.Gateways.QuotesGateway.Models
     public class HistoryQuoteInfo
     {
         public IEnumerable<long> t { get; set; }
-        public IEnumerable<decimal> c { get; set; }
-        public IEnumerable<decimal> o { get; set; }
-        public IEnumerable<decimal> h { get; set; }
-        public IEnumerable<decimal> l { get; set; }
+        public IEnumerable<double> c { get; set; }
+        public IEnumerable<double> o { get; set; }
+        public IEnumerable<double> h { get; set; }
+        public IEnumerable<double> l { get; set; }
         public IEnumerable<long> v { get; set; }
         public IEnumerable<DateTime> t2 { get; set; }
         public string s { get; set; }
